requirements.txt
Django==2.2.3
djangorestframework==3.10.2
pkg-resources==0.0.0
pytz==2019.2
sqlparse==0.3.0
